Support windows platform (#209)
* __PRETTY_FUNCTION__

* CASE_K

* uint

* remove not

* HALF_FLT_MAX

* struct init

* port utils

* better build pthread-win32

* port kernels

* port utils/gemm_test

* hide windows header

* port models

* port examples && triton_backend && unittests

* update build readme

* fix lint

* fix lint

* fix lint

* fix lint

* fix lint

* fix build

* fix build

* cmake version

* fix typos

* update ci

* port kernels/gemm_s_f16

* update ci

* fix ci

* use cudaStreamSynchronize instead of volatile check

* remove gettimeofday

* remove pthread-win32

* remove dirent.h

* update pre-commit

* update

* remove todo

* fix include

* fix build

* fix build

* fix build ci

* fix github action trigger

* update README

* fix linux-build ci

* remove windows folder

* fix lint

* update readme

+++ src/turbomind/models/llama/llama_kernels.cu
@@ -2,6 +2,7 @@
 
 #include "src/turbomind/kernels/decoder_masked_multihead_attention_utils.h"
 #include "src/turbomind/kernels/reduce_kernel_utils.cuh"
+#include "src/turbomind/macro.h"
 #include "src/turbomind/models/llama/llama_kernels.h"
 #include "src/turbomind/models/llama/llama_utils.h"
 #include "src/turbomind/utils/cuda_type_utils.cuh"
@@ -111,7 +112,7 @@ template<typename T>
 void invokeAddResidual(T* out, const T* in, int m, int n, cudaStream_t stream)
 {
     auto total = static_cast<size_t>(m) * n;
-    dim3 block(std::min(total, 1024UL));
+    dim3 block(std::min((unsigned long)total, 1024UL));
     dim3 grid((total + block.x - 1) / block.x);
 
     addResidual<<<grid, block, 0, stream>>>(out, in, total);
